Renamed files so they're not all index
